# api_old/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "跨链交易 API 服务",
        "title": "Fabric API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/{type}/query": {
            "post": {
                "description": "根据条件查询实体列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crud"
                ],
                "summary": "查询实体列表",
                "parameters": [
                    {
                        "enum": [
                            "orgs",
                            "parts",
                            "products"
                        ],
                        "type": "string",
                        "description": "实体类型",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "组织查询条件",
                        "name": "org",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Org"
                        }
                    },
                    {
                        "description": "零件查询条件",
                        "name": "part",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Part"
                        }
                    },
                    {
                        "description": "产品查询条件",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "产品列表\" SchemaExample(products)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Response"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "$ref": "#/definitions/models.Product"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.DID": {
            "type": "object",
            "properties": {
                "chainID": {
                    "type": "string"
                },
                "fragment": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "scheme": {
                    "type": "string"
                },
                "specificID": {
                    "type": "string"
                }
            }
        },
        "models.Org": {
            "description": "组织信息模型",
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "class": {
                    "description": "组织类型（如：ENTERPRISE）",
                    "type": "string",
                    "example": "ENTERPRISE"
                },
                "company_registration": {
                    "description": "工商注册号",
                    "type": "string"
                },
                "county": {
                    "description": "区县",
                    "type": "string"
                },
                "date_of_establishment": {
                    "description": "成立日期",
                    "type": "string",
                    "example": "2004-03-31"
                },
                "did": {
                    "description": "组织DID",
                    "type": "string",
                    "example": "did:org:66666666-0fac-4707-a41a-6a2674a720c8"
                },
                "email": {
                    "description": "电子邮箱",
                    "type": "string"
                },
                "enterprise_type": {
                    "description": "企业类型",
                    "type": "string"
                },
                "industry": {
                    "description": "所属行业",
                    "type": "string"
                },
                "juridical_person": {
                    "description": "法人代表",
                    "type": "string",
                    "example": "张美云"
                },
                "management_state": {
                    "description": "经营状态",
                    "type": "string",
                    "example": "开业"
                },
                "name": {
                    "description": "组织名称",
                    "type": "string",
                    "example": "扬州三圆核电电气有限公司"
                },
                "organization_code": {
                    "description": "组织机构代码",
                    "type": "string"
                },
                "phone_number": {
                    "description": "联系电话",
                    "type": "string"
                },
                "province": {
                    "description": "省份",
                    "type": "string"
                },
                "registered_capital": {
                    "description": "注册资本",
                    "type": "string",
                    "example": "7,188万(元)"
                },
                "social_credit_code": {
                    "description": "统一社会信用代码",
                    "type": "string"
                },
                "tax_number": {
                    "description": "纳税人识别号",
                    "type": "string"
                },
                "uuid": {
                    "description": "唯一标识符",
                    "type": "string",
                    "example": "66666666-0fac-4707-a41a-6a2674a720c8"
                },
                "website": {
                    "description": "网站",
                    "type": "string"
                }
            }
        },
        "models.Part": {
            "description": "零件信息模型",
            "type": "object",
            "properties": {
                "did": {
                    "description": "零件DID",
                    "type": "string",
                    "example": "did:part:66666666-a0e3-48dc-bff8-7eb7819b7a09"
                },
                "name": {
                    "description": "零件名称",
                    "type": "string",
                    "example": "围板"
                },
                "uuid": {
                    "description": "唯一标识符",
                    "type": "string",
                    "example": "66666666-a0e3-48dc-bff8-7eb7819b7a09"
                }
            }
        },
        "models.Product": {
            "description": "产品信息模型",
            "type": "object",
            "properties": {
                "did": {
                    "description": "零件关系DID",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DID"
                        }
                    ]
                },
                "end_id": {
                    "description": "零件ID",
                    "type": "string"
                },
                "end_label": {
                    "description": "零件名称",
                    "type": "string"
                },
                "fit_brand": {
                    "description": "适配品牌",
                    "type": "string"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                },
                "product_place": {
                    "description": "生产地",
                    "type": "string"
                },
                "relation_label": {
                    "description": "关系标签",
                    "type": "string"
                },
                "relation_name": {
                    "description": "关系名称",
                    "type": "string"
                },
                "start_id": {
                    "description": "供应商ID",
                    "type": "string"
                },
                "start_label": {
                    "description": "供应商名称",
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}